requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coverage==7.8.0
cryptography==44.0.2
defusedxml==0.7.1
Django==5.2
django-elasticsearch-dsl==8.0
django-redis==5.4.0
django-secure==1.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-nested-routers==0.94.1
elastic-transport==8.17.1
elasticsearch==8.18.0
elasticsearch-dsl==8.11.0
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.2.1
pluggy==1.5.0
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
urllib3==2.4.0
